working click on the background
